The beginning of a comprehensive, technical spec document
This document describes and specifies the IC Identity Provider from various angles and at various levels of abstraction, namely:

 * High level goals, requirements and usecase
 * Overview of the security and identity machinery, including the interplay of identities, keys, delegations etc.
 * Interface as used by client applications, i.e. the OAUTH-based spec
 * The interface of the IDP _backend_, i.e. describing its contract at the Candid layer, as used by its frontend
 * Important implementation notes about the IDP backend
 * Internal implementation notes about the IDP frontend
 * Notes about deployment

The goal is that
 * engineers can use this as a referenc to understand
the scopes and interplay of the component we are working on,
 * security review has a blueprint to start with
 * my understanding is double-checked by Björn and Jan and others

I am writing this as we go, so sometimes it might be laggin behind or ahead of the work. Please raise any questions you have, ideally as review comments attached to the relevant lines (so that we can “resolve” them). Suggestions welcome.

I’ll keep this on a PR until we are close to being done, so that the discussion is visible. Hence the draft state. I will not force-push, so you can catch up using the git or github history.