[IC-155] prune canister signatures after 10 min (#42)
This change adds code to linmit the lifetime of signatures to
approximately 10min.

The cost of removal is amortized across updates: each update call will
do some cleanup of the signature tree. This way the prunning can
always keep up with the addition of new signatures while keeping the
cost of pruning quite low (if there is nothing to do, the pruning
executes only a few instructions).

This implementation requires extra memory to maintain a priority
queue, 72 bytes per signature.