chore: fix package.json to generate JS and use CANISTER_ID env var (#93)
Since the candid bindings are generated AFTER the canister is built, but
here the frontend canister is built as part of the backend canister, the
bindings are creating a circular dependency.

This PR moves the generation of JS and TS bindings to npm scripts that
should be executed separately, and use the CANISTER_ID environment var
set by DFX to get the current CANISTER_ID.

Also did some cleanup on webpack config and tsconfig to make it simpler.