certified_map: add debug code checking unsafe memory usage
This change adds more debug code to the RbTree.

The new checks keep track of the pointers that are currently live and
crash if

  1. We delete a node that hasn't been allocated / has already been
     deleted (which would indicate a double free bug).

  2. Removing a node from the tree results in dangling pointers.

  3. There are some live pointers after the model test is
     completed (which would indicate a memory leak in delete()).

The new debug code didn't catch any bugs (because I'm almost sure
there are no bugs related to the correctness of memory allocation).
The tests do fail however if I intentionally introduce a memory leak /
double free (which is also detected by malloc itself).