Create screenshots with selenium
These screenshots are generated by running the selenium tests, e.g. using

    npm test -- -t Screenshot

That will override screenshots, but not delete files that are no longer
generated by the test; in that case, run `rm -f screenshots/*.png` before
running this script.

We commit these screenshots so that changes to the UI can be reviewed nicely in
the PR.

At some point we might
 * make CI fail if the screenshots are not up-to-date
   (assuming they can be made reproducible on different machines, or else
   diffed liberaly)
 * add a bot that updates these automatically