compute asset hashes at build time
This change introduces a custom cargo build script that generates
assets module with pre-computed asset hashes.

This removes the need to compute hashes on init/upgrade, which takes a
significant amout of cycles.  Backend tests say it all:

Before:

    Tests
    without upgrade
        installs:                                 OK (28.33s)
        installs and upgrade:                     OK (53.17s)
        register with wrong user fails:           OK (28.83s)
        ...
    All 48 tests passed (655.16s)

After:

    Tests
    without upgrade
        installs:                                 OK (0.88s)
        installs and upgrade:                     OK (1.65s)
        register with wrong user fails:           OK (1.35s)
        ...
    All 48 tests passed (49.85s)

The script can later be extended to run `npm run build` automatically
to eliminate the chance of accidentally building a canister with
outdated assets if `cargo build` is used directly.