optimization: use Cow<'static, Bytes> for HttpResponse body
This change speeds up asset serving a bit by avoiding an unnecessary
copy when HttpResponse is constructed.

Asset bytes are static and can be served directly from pre-initialized
memory, which is what this change is trying to achieve.

The instruction consumption for the biggest asset went down from ~13M
instructions to ~9M. Ideally, asset serving should be almost
independent of the asset size, there is probably still an extra copy
somewhere in CDK.