Client auth flow: pass maxTimeToLive (FOLLOW-136)
We were not passing `maxTimeToLive` through to the backend. This fixes
this.

It also refactors the selenium test suite mildly and adds tests that
exercise setting the default maxTimeToLive, 1min, 1day and 1 month.

This needs a released update in `auth-client` for the tests to work, so
we’ll have to wait for https://github.com/dfinity/agent-js/pull/417