Spec: The “add device“ link is `#device=`, not `#add_device`
It used to be `#device` in implementnation and spec, but at some point I
changed it to `#add_device`, because it seemed to be cleare in its
intend. But I guess we never actually implemented that… so let’s not
bother, and get our docs up to date before people are confused.