Pin verison of ic-cdk-optimizer
it is important to make our build as hermetic as possible, so that
different people, even at different times, can reproduce the `.wasm`
that we deployed. The build infrastructure we use (Docker) is not fully
hermetic by default, so we need to root out possible variations
manually.

This pins the version of ic-cdk-optimizer.