Re-enable whoami check without ic-ref (#425)
This is a horrible hack. The 'whoami' canister does not respond
when running with the emulator, meaning we had to disable
one check
in our test suite. It is taking longer than expected to
figure out
why, so for the time being we make sure to at least run the
full
test suite when _not_ using the emulator.