Pause for 0.5s before taking screenshot
We have some flakiness in screenshots that can't really be explained.
The best hypothesis thus far is that the screenshot is taken after the
first rendering pass, but before it has settled on exactly how to render
everything, leading to a few (8) pixels having different colors. If this
doesn't solve the issue, then we'll need to hack and change how screenshots are
committed, and only commit if the image was modified beyond some
threshold.