backend-tests: Read Candid types from .did file
previously, they had to be copied because of a infelicity in
haskell-candid, with this this should be fixed.

I am unsure about the implications on compile time, this is using lots
of `row-types`. If it is a problem, might be worth trying to upgrade
GHC. `ic-hs` is at 8.10, so we could switch too.