Remove WebAuthnIdentity from iiConnection (#536)
This changes `IIConnection` to never actually use `WebAuthnIdentity`,
which gives us more flexibility. This is the first step before we add
the switch for jailbreaking II with the dummy, `\NUL`-byte-private-key
identity.

This also documents some of the Identity lookup code and changes
`MultiWebAuthnIdentity` to actually fallback to its "actual identity" a
bit more safely -- and explicitly -- after it's been "discovered".