Remove Init arguments
The init arguments allow us to specify different anchor ranges when
deploying the II canister, the goal being that anchors can be sharded
across different instances.

We don't actually support that, and I don't think those arguments have
ever been used, but they are a real hassle when deploying (due to the
dependency on didc) and they are very confusing. Moreover, if/when we
start sharding II, it's likely things will have changed and that the
arguments won't actually be usable as is.

I've left a comment in the code pointing the reader to the correct
commit if they start looking for ways to set new `id_range_{lo,hi}`
values.