Clean up HACKING.md, CONTRIBUTING.md
When splitting part of the README.md into a new CONTRIBUTING.md
document I forgot we already had a
`.github/CONTRIBUTING.md`. This moves `.github/CONTRIBUTING.md` to the
top-level, and moves hacking instructions to HACKING.md. This also
rewords some of the README so that it makes sense in the new structure.