Show banner if origin isn't identity.ic0.ap (#618)
We should make sure that unsuspecting users do not land on testing
builds of II. While we can't ensure that, we can at least let them know
that something is wrong. So far we've displayed a banner if the build
included features; this adds a banner if the origin is not
identity.ic0.app.