Record size of wasm in git notes
This adds a CI step that computes the size of
`internet_identity_production.wasm` and records it in a git note. The
size is also checked against the maximum canister size. The git note is
created and pushed only on the main branch.

The logic is pulled into a GitHub action because the `canister-tests`
workflow file is getting gigantic.