Show warnings during recovery and skip validation
This addresses some long standing issues related to recovery phrase
verification/validation and BIP39, and (in my opinion) improves UX a
bit.

The two important changes are:

* Warnings are shown during recovery phrase input: these includes
  warning about words that aren't recognized, as well as other not so
  easy to spot typing errors.
* Invalid recovery phrase won't prevent user from submitting: this is
  pretty much the first sentence on [bitcoinjs' bip39
  repo](https://github.com/bitcoinjs/bip39#reminder-for-developers);
  instead, it is only a warning.

In a bit more details, the `inputSeedPhrase` page (where the user inputs
their recovery phrase) now shows the following warnings as the phrase is
being typed:

* Unexpected characters were entered (we're expecting letters and
  numbers only)
* Unexpected words were entered (we're expecting words from a list)
* Recovery phrase was generated for another anchor (already existed)
* Incorrect word count (we're expecting 24 words)
* Recovery phrase is not bip39-valid
* Double whitespaces were found (which _will_ prevent the user from
  recovering their anchor)

One thing to note is that, aside from trimming whitespaces at the
beginning and and of the recovery phrase, the implementation does _not_
modify the phrase input by the user, and will _not_ prevent the user
from submitting a phrase deemed invalid. As far as we know we've only
issued properly formatted, english bip-39 phrases but we can't be
certain.

This also adds a bunch of tests for `mnemonic.ts` which now also
regroups most of the operation on the recovery phrase (and fixes a bug
or two spotted by tests).