Converting II-Spec to Markdown
Hi II folks! 
With the new dev portal, we are converting the docs from adoc to md. Of course we would like the reference documentation to live in their home repo. Here I'm pushing a markdown version of the current II Spec. This is the file that will be pulled into the portal repo, so it would be great if further updates could be made in the md version. 
If you've any questions or comments, feel free to shout. Thanks for the patience as we transition to the new docs :)