Pass RUSTFLAGS in dependency build
This instructs the Dockerfile to use the official build script, which
now takes an extra option for building only rust dependencies.

When introducing `--remap-path-prefix` I forgot to set the build flag
during the dependency build, the actual build in Docker would rebuild
most of the crates. Now the Dockerfile uses the build file as well.

I've also added a new script, `./scripts/build`, for ease of use.