Add scheduled release build checks
This adds "clean" (i.e. from scratch) daily builds using a checkout of
the latest release. We are notified of an error if the builds fail or
if the resulting sha256 hash of the production asset (on Ubuntu) does
not match the sha256 hash of the release assets.

This helps us ensure that our builds are reproducible over time, meaning
not just reproducible at the time of release, so that users can confirm
that the Wasm module running in production was built by the release
code.