Adds rust tests for the registration flow
The registration flow consistes of two canister calls: create_challenge and register.
As part of testing these canister calls the inteface of api::create_challenge has been changed to return a Result<Challenge, CallError> to make error behaviour testable.
Related to registration, this PR includes tests related to the intialization of the user range (ported from haskell).
The test for the user range widening has been moved to the http test module (together with other metrics) since it is unrelated to user registration.