Add option to protect recovery phrases (#721)
This adds an option for users to make their recovery phrases more secure by marking them as "protected". When a device is marked as "protected", the user must be authenticated with the same device to modify/remove it.

As far as the webapp is concerned, only recovery phrases can be protected. The canister code was changed to allow general mutation (i.e. modifying devices atop removing them) which is guarded in the case of protected devices. While in general the canister code does not make any assumption regarding whether a protected device is a recovery phrase or not, for now it will prevent marking a device as protected unless it is a recovery phrase.

Here are some important changes:
* A new canister method, update, was introduced
* The DeviceData was updated to include a protection_type. The internal representation is backwards compatible to ensure old anchors can still be read from stable memory. (thanks @olsemeno)
* A new "device settings" page was introduced that replaces the removal button. Devices can be removed from that settings page.
* The webapp will prompt the user for the recovery phrase in case the phrase is protected (thanks @lastmjs)
* The canister test API module was updated to add a compatibility layer so that it can read device info both from old and new code (since the DeviceData candid representation changed)


Co-authored-by: olsemeno <sendtosemenov@gmail.com>
Co-authored-by: Jordan Last <jordan.michael.last@gmail.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Frederik Rothenberger <frederik.rothenberger@dfinity.org>