Remove headless option for selenium tests and improves information on test failures (#745)
The reason is that the clipboard permission is denied in headless mode. To keep screenshot sizes the same, chrome is run in mobile emulation mode for the mobile tests.
The log output on selenium test failures has been optimized to be more concise and provide additional information in files.
This was done, because the clipboard issue would have been found much quicker if the error output had been better.