Don't symlink to candid file from demo
Before this the candid file was symlinked from `using-dev-build`. We do
however suggest to copy the directory out of the repo for users to start
their II-enabled projects, which would lead to a broken symlink.

Instead of suggesting to copy the target file as well, we simply put the
candid file in the release (as also suggested in https://github.com/dfinity/internet-identity/issues/651) and download it
like we download the wasm file.