Add version_info to canister and autodeploy
This adds a `version_info` field to the init args that is used to
bake in some build-time information, which is later shown in the browser
console. This helps track which version was installed.

This also adds a deploy step to the CI runs on the `main` branch that
deploys to our mainnet test canister.