Read canisterId lazily
This ensures that the canister ID is read lazily, meaning devServer
won't crash if there is no local dfx state. The error will still be
shown but when performing the first (if any) request that requires
routing to the canister. This means that requests that do not need to be
routed to the canister won't crash if there is not local dfx state.