Add a showcase devServer
This adds a new webpack configuration and `npm run start` script that
serves a showcase of some of II's pages (right now only the simplest
ones).

This should greatly improve productivity when working on HTML and CSS
(and some JS) changes, since one does not have to go through II flows to
land on a particular page.