Skip canister_tests build if necessary (#835)
This greatly improves cache reuse for canister tests build on GHA.

This adds a step that sets the mtime of all files relevant in the
canister tests builds to the date they were last edited, according to
the git history. This means that in most cases (e.g. when the canister
tests' code is not changed) the build is skipped entirely (though tests
are still run).

In order to achieve this the checkout step is changed to fetch the
entire git history, which takes 4s instead of 3s for a time gain of
about 5mn.