Disable cursor before taking screenshots (#871)
This disables the cursors before taking a screenshot and reenables it
afterwards. The blinking cursor was causing issues in screenshots, as
screenshots would look different depending on timing.

There were issues on page with no apparent blinking cursor, but
disabling the cursor seems to have updated these pages too. I'm assuming
this means the problem was fixed there as well although I'm not sure
what the exact problem was.