Show chasm instead of tooltip for alternative origin
This shows a "chasm" (which splits the card open to reveal details)
instead of a tooltip on the authenticate screen.

This makes the screen a little sleeker and ensures mobile users can
display the details too. The wording is changed a bit to make the
information more useful and understandable.