Shift anchors in stable memory and double size
This PR implements the first part of the stable memory migration, which
is shifting the anchors in stable memory and doubling their size.

The migration can be configured via init arguments and will shift
the configured amount of anchors on each stable memory write.
This way we can continually shift anchors without having to implement
heartbeat. The migration can also be paused at any point.

The second part of the migration (not included in this PR) will
change the candid schema in place for each anchor.