Store multiple anchors in local storage
This updates our local storage handling to actually store up to 10
anchors, which we then all display on authentication pages.

UX note: We keep "infinitely" many anchors because we want to visually
discourage users from creating too many anchors (but really we keep
only 10 for practical reasons); we also keep as many anchors as
possible because this is the only place where the user will see all
their anchors.

Implementation note: we make the anchor reading/parsing as
non-destructive as possible in case we e.g. add a field in the future
and then need to rollback for some reason but want to keep the field.