Serve static pages statically
This updates our webpack config to pre-render the FAQ and about pages,
which are then served as HTML from the canister (and then hydtated in
the browser). This means that `/about` and `/faq` can be browsed (and
more importantly scraped/indexed) without JavaScript. A side-effect is
faster page load.

The process for generating static pages is a little manual right now,
though we may automate it/abstract it if we decide to have more static
pages:
1. Use `lit-ssr` to render a string version of the page during the
  webpack build
1. Inject that content in the HTML page served by the canister
1. Hydrate the page before the first `render()`