Update the metrics encoder
The archive requires more features of the metrics encoder. This PR
adds the latest changes from the IC metrics encoder to this repo.

To avoid having to do such PRs in the future, I am also in the process
of opensourcing the metrics encoder as a standalone crate. This PR
just bridges the gap until we have it ready.