Temporarily disable screenshots
The latest version of chrome triggered a bug so that
screenshots on mobile now have the wrong aspect ratio and
produce flaky results.

Upgrading webdriverio to version 8 would fix the issue,
however this then breaks compatibility with jest.

Since it is a bigger change to make the screenshot infrastructure
work again, this PR disables the screenshots (temporarily)
so that we can deal with fixing the infrastructure properly at a later stage.