Refactoring to remove redundant optionals from internal device struct
This PR is in preparation of the upcoming candid schema migration of the
devices stored in stable memory. The schema changes will go from `Vec<DeviceDataInternal>`
to `Anchor`. In order to keep the diff for the migration code change small
this PR confines the type `DeviceDataInternal` to the `storage` module.