Remove initialization of salt when it's not required
This PR removes the calls to init_salt from anchor operations
because they have nothing to do with the salt. This allows many
anchor operations to become sync, thus simplifying the code.

Additionally it moves some interleaved calls to `prune_expired_signatures`
to either the beginning or the start of the respective functions.