Rename UserNumber to AnchorNumber
This PR only moves and renames things. It cleans out the `Anchor` alias
in the interface crate (because it clashes with the `Anchor` struct from the state module).
To do that, `UserNumber` and `Anchor` type aliases got unified to `AnchorNumber`.
Additionally, in some modules the imports are reorganized to reduce clutter.