Remove initialization of salt when it's not required (#1095)
* Remove initialization of salt when it's not required

This PR removes the calls to init_salt from anchor operations
because they have nothing to do with the salt. This allows many
anchor operations to become sync, thus simplifying the code.

Additionally it moves some interleaved calls to `prune_expired_signatures`
to either the beginning or the start of the respective functions.

* Initialize the salt explicitly in tests

Some tests require initializing the salt explicitly because they
call `get_principal` without preparing a single delegation first.
Note that in practice this will not be an issue, as the II front-end
does not use `get_principal`.