Refactor persistent state in preparation for archive pull integration
This PR refactors the archive state handling and prepares the persistent
state for the additional configuration parameters required when the archive
integration is switch to pull entries from II (instead of II pushing them to
the archive).

This PR introduces changes that require a persistent state migration which
can be triggered by setting the corresponding flag in the II init arguments.