Refactor CAPTCHA and improve UX
This refactors the captcha code a bit and changes the UX to clarify what
is expected of the user.

In particular, the captcha page is now more reactive. `Chan`s are used,
which are `asyncReplace`d in the template on state changes (see code for
actual states).

In terms of UX, a new `retry` button is introduced and the page prompts
the user to retry on failure.