Add version to non-docker builds
This updates the `./scripts/build` to set `II_VERSION` using
`./scripts/version`, if the environment variable isn't already set.

This also slightly changes the version format to _always_ print a value
for the "dirty" line (either `"dirty"` or `"clean"`) since it's
otherwise quite difficult to read the output of the script in bash (as
bash will strip trailing empty new lines in command substitution).