Change recovery picker depending on context
This updates the component we use to pick a new recovery method to show
different text depending on whether we're adding a recovery during
sign in/sign up or from the "manage" page.

The implementation shoehorns the changes in the current project
structure which should be revised a little. In particular, the
`chooseRecoveryMethod` and `setupRecovery` modules should be moved to
`components/`. This will be done separately.