Passthrough anchor to recover and add screens (#1252)
This skips asking the user for their anchor when adding a device or
recovering an anchor, _if_ the user has already typed in an anchor
number. This simplifies the flow and is probably what the user expects.

Some of the subsequent screens (e.g. prompting for device alias or
recovery phrase) are adapted to show the (potentially inferred) anchor
number for clarity.

The templates are extracted from `authnPages` to allow for testing.