Allow 0.0.0.0 for local development
This ensures the local API is used when developing on `0.0.0.0` instead of
trying to use the `icp-api.io` API.

Recent changes introduced using `icp-api.io` for non local development.
The heuristics for figuring out whether yes or no the page was served
locally didn't include `0.0.0.0`, which (while uncommon) is a viable
host for local development.