Don't mention alternative origin of self (#1258)
This avoids showing the "shared identity" prompt for alternative origins
if the derivation origin is the origin itself.

This may happen when a dapp is served on multiple domains. Technically
the dapp shouldn't have to specify a derivation origin when it's being
served on the derivation origin itself, but just in case, we don't show
the message.