Don't show menu on no settings
This fixes the manage template to show no menu when no settings are
available.

The previous version used boolean shortcutting to decide whether or not
to display the menu (`foo && menu`) which unfortunately defaults to show
`false` in case no settings are provided. This fixes it by returning
`undefined` instead, which `lit-html` interprets as showing nothing.