Deduplicate alias prompt
There were two pages prompting for a device alias. This merges them into
a single component. The `promptDeviceAlias` page was favored, as it was
tested and had received extra CSS love.

The main difference between the two pages was the title, which is no
passed as an argument. Some orphan comment about alias validation was
moved to `validateAlias`.