Fix /faq redirect
This PR fixes the /faq link redirect by setting the upgrade to update flag on the redirect
and allowing http_requests to be sent as update calls.

The redirects can be certified for query calls with the V2 certification at which point
the support to receive http assets via update call can be dropped again.