Make manage flows explicit
The `renderManage` page is the starting point for many flows, but all
these flows also lead back to it. This was somewhat implicit.

These changes make the "infinite loop" nature of the management page
explicit and updates all flows to _not_ decide what to do next; instead
this is left to the caller (often `renderManage`). This makes the flows
more modular and clarifies the return values.

This made an issue apparent in `pollForTentativeDevice`, where the
loader was never exited; now fixed.