Remove lookup from tests (#1293)
* Remove lookup from tests

`lookup` was recently deprecated. As a consequence, integration tests should not
rely on the usage of `lookup`. This PR refactors the tests to not use it anymore,
except for one test that ensures that `lookup` keeps working as intended.

* Add test for fallible type conversion

* Fix merge error

The automatic merge did not consider the rename of WebauthnCredential
to WebAuthnCredential.