Show recovery words table
This updates the page showing the recovery phrase to be more user
friendly.

The new component (designed by @meodai, adapted here from the
styleguide) displays each word separately, similarly to what most
wallets do.

The tests are updated to use the preferred storage method, which is the
"copy" button; previously tests copied the value from the `output`
element, which should but may not have matched the clipboard content.

Some workarounds are needed to make the clipboard work in the browser
tests; see code for details.