Add recovery phrase confirmation
This adds a screen asking the user to confirm (re-input) part of their
recovery phrase.

This only covers the logic & copy of the screen; the design is implemented
separately.

The confirmation is done by presenting the user with their phrase with
some words removed. The words need to be re-typed. The user may only
continue once all words have be re-input correctly. This is done by
storing the expected word as `data-expected` and regularly checking all
words against their `data-expected` fields. Per-word feedback is also
given to show which words have been input correctly and which still need
attention.