gzip the production asset wasm (#1330)
* gzip the production asset wasm

This PR introduces gzipping the production wasm and publishing it
on the releases page.
It is a prerequisite to dropping ic-cdk-optimizer in favor of ic-wasm,
which does not shrink the wasm as much but offers convenient editing
of the wasm metadata.
This in turn will prepare II pipeline for upcoming requirements with regards
to the canister metadata, such as providing the candid interface or metadata
for dfx pull.

The release build check / clean build checks are not yet switched because
it requires at least one release that provides gzipped assets.

* Use longform command line args for gzip