Add domain specific active anchor statistics
This PR introduces additional statistics about active anchors
based on the domain of the device that was used for the activity.

The stats are collected only for the well-known II domains
(and the combinations of both).

Caveat: the stats will be inaccurate for identity.ic0.app because
most devices there still do not have domain information attached.
This issue will be addressed in a follow-up PR, where domain
information will be added on activity if it is missing.