Switch from ic-cdk-optimizer to ic-wasm
This PR introduces `ic-wasm` in place of the deprecated `ic-cdk-optimizer`.
`ic-wasm` will allow to modify the medatada sections of the generated
wasm files. However, `ic-wasm shrink` does not achieve the same
level of size reduction as `ic-cdk-optimizer` did. This is why
the production asset will be switched to the gzipped wasm in an upcoming
PR.