Update device origin on usage instead of key type (#1384)
* Update device origin on usage instead of key type

This PR replaces the hack to update the device key type
(now obsolete due to 1381) with a hack to update the origin
instead.

We need origin information for existing devices in
order to accurately report user activity by II domain.

Since we already add origin information to new devices, this hack
can be removed in a few weeks, when most active II users with old
devices have had some activity.

* Add type guard for device.origin

Co-authored-by: Nicolas Mattia <nicolas.mattia@dfinity.org>

* Fix formatting

---------

Co-authored-by: Nicolas Mattia <nicolas.mattia@dfinity.org>