Enable dynamic certified assets (#1391)
This PR enables dynamic certified assets by creating a copy on the
heap and another copy when assembling the http response. While
less elegant, the data is small enough to not justify a more
complex solution.

This is a prerequisite to have the showcase dapp list served
as a certified json file hosted on II. It will also allow
serving assets from stable memory in the future.