Add add-device link
This changes the "add device" flow slightly (when starting from an
existing device) to show a link to be opened on the new device.

The link looks like this: `HOST/?action=add-device&anchor=1234`

This is then interpreted by the new device as a cue to be added
tentatively to anchor `1234`.