Extract index message and copy
This extracts a `showMessage` component/function from the index and
extracts the copy too.

The new `showMessage` is also reused in other places that need to
display a centered message, and the `data-role` is fixed (in index.ts,
the attribute had been copy/pasted but was incorrectly suggesting
some authorization was successful).