Add modals without form (#1420)
* Add modals without form

Before this, modals could not be created without forms. This allows
creating modals that can be closed though do not necessarily contain
forms.

The previous default behavior (a modal with form) is moved to
`formModal`. I'm not sure we need it but leaving it because it's used
in the style guide.

* Move comment