Remove static copy
Static versions of copy were used in places where a `TemplateResult`
couldn't get inlined in an another `TemplateResult` (for instance, in
HTML attribute values). On the other hand, `DirectiveResult`s can be
inlined nicely, so this changes the `DynamicCopy` to produce
`DirectiveResult`s instead of `TemplateResult`s.

A new helper type is introduced (`TemplateElement`) for all "elements"
that can be inserted in templates (strings, etc).