Introduce temporary keys (#1171)
* Introduce temporary keys

A temporary key can be used in lieu
of the device but has a short expiration time. These keys are used as a workaround for WebAuthn
needing two users interactions: one for "create" and one for "sign". So instead we only "create"
and instead authenticate the user with a temporary key for their first visit.

Note: we link the temporary keys to a device (as opposed to an anchor) so that we can make sure
the temporary key is dropped if the device itself is removed. This ensures the temporary key is
no more powerful than the device, i.e. if the user decides to remove the device right after
registration, then the the temporary key cannot be used to authenticate (similary to how the
browser session key pair cannot be used to authenticate if the actual delegated WebAuthn device is
removed)

* Fix clippy complaint

* Fix formatting issue

* Wip

* Refactor temp keys and add anchor binding

Refactor the temp keys implementation:
* move temp keys code into its own module
* make expiration clean-up amortized
* bind temp keys to a specific anchor

* Prevent temp key being equal to device key

* Add tests

* Fix clippy complaints

* Trigger CI

* Add metrics

* Implement review feedback

Changes:
* switch amortized clean-up to use a deque
* tie clean-up and validity checking together
* remove unnecessary error type
* refactor auth check to prevent mismatching anchor and anchor number

* Add comment about expirations sorting

---------

Co-authored-by: Frederik Rothenberger <frederik.rothenberger@dfinity.org>