Refactor canister test API to work with borrows (#1451)
* Refactor canister test API to work with borrows

This is refactoring to make the canister test APIs more ergonomic:
In particular, wherever possible the API no longer takes ownership
of the arguments. `String` args are changed to work with `&str` instead.

This paves the way to slowly move away from the test helpers in `framework`
and use more appropriate local helpers instead.

* Revert inlining front-end hostnames