Add device metadata (#1447)
* Add device meta data

This PR adds a flexible a flexible meta data map to each device.
It can be used to more flexibly store data within the II canister.
Currently the meta_data map supports String, Blob and nested maps.

The top-level metadata key is not private (as changed keys are recorded
in the archive). However, all values (and nested maps) require
authentication to access.

* Rename meta data -> metadata

* Implement review feedback

Also add a notice to account for the planned migration.

* Fix archive stable memory compatibility

* Enforce reserved keys