Add 'Save Passkey' page
This introduces a new page, "Save Passkey", that prompts the user for
creating a WebAuthn identity, while giving some information on what a
"passkey" is.

This also extracts the "rendering" bits of `constructIdentity`, such
that `constructIdentity` now only focuses on the WebAuthn interaction,
and the caller is responsible for e.g. showing a loading screen.