Infer authenticator aliases
This introduces heuristics to give default aliases to authenticators.
The alias prompt is removed from the flows (the user will have to
explicitly rename the device if the alias is not ok).

The UAParser module is lazily loaded to make sure existing users don't
incur the cost of an increased bundle size.