Preload Identity card background (#1567)
* Preload Identity card background

The "Internet Identity" card background is an image that must be loaded
from the canister, which can take a few seconds. This ensures that when
used in the identity creation flow, the browser can already start
fetching the image as the user is dealing with the CAPTCHA
(optimistically assume the user will succeed, at very little cost if
the user does not actually succeed).

* Fix typo