Remove outdated documentation
These guides are no longer hosted on internetcomputer.org from
this repo. Instead there is a copy in the portal repo now.

The guides will be reworked and updated by the support team, so
there is no need for us to keep them around anymore.

Also antora is no longer in use, so all configuration files for that
tool can be removed too.